BUG-5141: Failed to send Update Message on reconfigure Peer
Clean up when reconfigurin Peer as we do when we reconnect peer.
Update Log levels on LocRibWriter for skipped change.
Making Debug logs more clear.

+++ bgp/rib-impl/src/main/java/org/opendaylight/protocol/bgp/rib/impl/AdjRibInWriter.java
@@ -244,13 +244,15 @@ final class AdjRibInWriter {
     }
 
     void removePeer() {
-        final DOMDataWriteTransaction tx = this.chain.newWriteOnlyTransaction();
-        tx.delete(LogicalDatastoreType.OPERATIONAL, this.peerPath);
+        if(this.peerPath != null) {
+            final DOMDataWriteTransaction tx = this.chain.newWriteOnlyTransaction();
+            tx.delete(LogicalDatastoreType.OPERATIONAL, this.peerPath);
 
-        try {
-            tx.submit().checkedGet();
-        } catch (final TransactionCommitFailedException e) {
-            LOG.debug("Failed to remove Peer {}", this.peerPath, e);
+            try {
+                tx.submit().checkedGet();
+            } catch (final TransactionCommitFailedException e) {
+                LOG.debug("Failed to remove Peer {}", this.peerPath, e);
+            }
         }
     }
